darwin.nettools: init
darwin.nettools is a fake derivation to provide some network commands
to derivations that expect the Linux-only net-tools. This is for
aliasing purposes only so only $out/bin/* files are provided. This is
the list of commands:

- arp
- route
- ifconfig
- netstat
- hostname

They are not guaranteed to have the same functionality as net-tools
but it should be enough to trick most applications. These commands are
provided by Linux-only net-tools but NOT available in Darwin nettools:

- rarp
- plipconfig
- slattach
- mii-tool
- iptunnel
- ipmaddr

No compatible versions of these commands exist for OS X. Perhaps there
are equivalent ways to accomplish their functions. In the future, we
should attempt to package wrappers for these commands. See
"iproute2mac" for more info[1]. Other packages that may need to be
aliased are: iproute, iptables, procps, and ethtool.

[1]: https://github.com/brona/iproute2mac