stdenv: add shell to default build inputs
This puts ‘bash’ in HOST_PATH. That fixes patch-shebangs behavior,
restoring it to the old behavior (where Bash could wind up in build
outputs even when not specified). Bash should be small enough for this
to not have any significant issues. Needs testing to make sure this
doesn’t break stdenv bootstrapping.

Also, I’m not sure if ‘shell’ in this case is the builder’s shell or
the target’s shell. Most likely it is the builder’s shell which will
cause issues for cross compilation (wrong shell ends up in outputs).
Need to figure out how to do this correctly with bootstrapping
concerns. Unfortunately this is probably the first time we have had to
know the target’s bash ahead of time.

/cc @copumpkin @Ericson2314