llvmPackages: keep darwin at llvm5
My previous attempts to bump the Darwin stdenv version didn't
seem to work, but AFAIK doing so should be straightforward
for anyone with a Darwin builder and the interest in tackling this.

While not the preferred solution, adding a conditional lets us move on,
as we missed 6 entirely due to no one getting around to resolving the
Darwin bootstrap accordingly.