treewide: remove unstable packages
There’s nothing wrong with unstable packages when they are maintained
and frequently updated. However, when they become out-of-date, as many
do, it is usually best to just get rid of them as the stable version
is become newer than the unstable version. This removes any packages
called "unstable" that have not been updated in over 1 year. Affected
packages include:

- ingen-unstable
- cataract-unstable
- isyncUnstable
- lv2Unstable
- bspwm-unstable
- sxhkd-unstable
- minisatUnstable
- dosbox-unstable
- xdeltaUnstable