nixUnstable: disable doc generation in cross
Unfortuantely, nix 2.4+ relies on running the built nix executable to
generate documentation. We can’t do that when cross-compiling, so we
need to skip that step for now. In the future, an emulator could be
used to run the built Nix binary.

Fixes #106415

(cherry picked from commit 85f43b1834bc347a3220b183c672ed5a898da465)