gcc: update mcf thread support to latest version
This contains the gcc10 patchset, although it’s the same as the gcc9
one which is the same as the gcc8 one. It looks like there have been
some minor changes however.

Removes mfcgthreads-patches-repo.nix so that the full URL & the hash
are closer together. This makes it easier to update values.